Reformat the tests
