--- A02/tests.h
@@ -1,11 +1,0 @@
-#ifndef __TESTS_H__
-#define __TESTS_H__
-
-#define TRUE 1
-#define FALSE 0
-
-extern int test_iconst(void);
-extern int test_chrconst(void);
-extern int test_keyw(void);
-
-#endif
